backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
scikit-image==0.21.0
exifread==3.0.0